webuihelper/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-constance==2.9.1
django-picklefield==3.1
idna==3.4
psycopg2-binary==2.9.7
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.4
